+++ src/pages/AddressPage/tabs/transactions/columns/transactions.tsx
@@ -1,100 +1,9 @@
-import {Box, ColumnConfig, Text, Tip, DropButton, TextInput, Button} from "grommet";
-import {Close, Filter as FilterIcon} from 'grommet-icons'
-import React, {useState} from "react";
+import {Box, ColumnConfig, Text, Tip} from "grommet";
+import React from "react";
 import { RelatedTransaction } from "../../../../../types";
 import { Address, DateTime, ONEValue, TipContent } from "../../../../../components/ui";
 import { TransactionAddress, TransferDirectionMarker, TxMethod } from "./common";
-import styled from "styled-components";
-
-interface ColumnFilterProps {
-  initialValue?: string
-  onApply: (value: string) => void
-}
-
-const FilterDropButton = styled(DropButton)`
-  border: none;
-  background-color: rgba(119,131,143,.1);
-  transition: background-color 0.2s;
-  padding: 4px;
-  border-radius: 4px;
-
-  &:hover, &:active {
-    background-color: #77838f;
-    box-shadow: 0 4px 11px rgb(119 131 143 / 35%);
-  }
-`
-
-const ColumnFilter = (props: ColumnFilterProps) => {
-  const { initialValue = '', onApply } = props
-
-  const [value, setValue] = useState(initialValue)
-  const [isValueApplied, setValueApplied] = useState(!!initialValue)
-  const [errorMsg, setErrorMsg] = useState('')
-
-  const validateValue = (v: string) => {
-    if(v.length > 0) {
-      if(!v.startsWith('0x')) {
-        return 'Address should start with "0x..."'
-      }
-      if(v.length != 42) {
-        return 'Address should contains 42 characters'
-      }
-    }
-    return ''
-  }
-
-  const applyButtonLabel = <Box direction={'row'} justify={'center'}>
-    <FilterIcon size={'12px'} color={'text'} />
-    <Text color={'text'} size={'12px'}>Filter</Text>
-  </Box>
-
-  const onApplyClicked = () => {
-    const err = validateValue(value)
-    if(err) {
-      setErrorMsg(err)
-    } else {
-      onApply(value)
-      setValueApplied(true)
-      setErrorMsg('')
-    }
-  }
-  const onClearClicked = () => {
-    setValue('')
-    onApply('')
-    setValueApplied(false)
-
-  }
-
-  return <FilterDropButton
-      label={<Box direction={'row'}>
-        <FilterIcon size={'16px'} color={'text'} />
-        {value && isValueApplied &&
-            <Box direction={'row'} gap={'8px'} justify={'between'} align={'center'}>
-              <Text size={'12px'} color={'text'}>{value.slice(0, 5)}...{value.slice(-3)}</Text>
-              <Close size={'12px'} onClick={onClearClicked} />
-            </Box>
-        }
-      </Box>}
-      dropContent={
-        <Box pad="small" background="text" style={{ borderRadius: '2px' }}>
-          <TextInput
-              placeholder={'Search by address e.g. 0x..'}
-              value={value}
-              size={'xsmall'}
-              onChange={(e) => setValue(e.target.value)}
-          />
-          {errorMsg && <Box margin={{ top: 'xsmall' }}>
-            <Text size={'xsmall'}>{errorMsg}</Text>
-          </Box>}
-          <Box direction={'row'} margin={{ top: 'small' }}>
-            <Button primary label={applyButtonLabel} onClick={onApplyClicked} />
-            <Button label={'Clear'} onClick={onClearClicked} />
-          </Box>
-        </Box>
-      }
-      dropProps={{ margin: { top: '32px' } }}
-  />
-}
+import {ColumnFilter} from "../ColumnFilter";
 
 interface ColumnFilters {
   [property: string]: {
@@ -237,7 +146,11 @@ export function getColumns(
           >
             To
           </Text>
-          {columnFilters && columnFilters['to'] && <ColumnFilter initialValue={columnFilters['to'].value} onApply={columnFilters['to'].onApply} />}
+          {columnFilters && columnFilters['to'] &&
+              <ColumnFilter
+                  initialValue={columnFilters['to'].value}
+                  onApply={columnFilters['to'].onApply}
+              />}
         </Box>
       ),
       render: (data: RelatedTransaction) => <TransactionAddress id={id} address={data.to} width={'180px'} />,
